New Files and Updates
history.py is the file for the cronjob. Should work exactly as is currently. Needs profile_id.csv to pull the profile_id's from and of course scraper.py for the functions. In cronjob just run 'python history.py' and it'll grab all UTR history of the last year from every player in the profile_id.csv file one time through.