Prediction File
Given 2 player names that are in the utr_history database (data collected and updated from cronjob), file will predict the winner, percentage of games won, and the score of the match. atp_utr_tennis_matches.csv is used for model training. Need to look into saving the model so it doesn't have to be retrained every time. Also want to write a second web scraper script that will add to the training data.